refactor: mock login 로직 및 파일 이름 등 수정 [TRL-144]
