+++ src/mocks/handlers.ts
@@ -1,24 +1,22 @@
 import { rest } from 'msw';
 
+type OauthServerKey = 'google' | 'kakao' | 'naver';
+
 const sleep = (ms: number) =>
   new Promise(r => {
     setTimeout(r, ms);
   });
 
-let error = true;
-
-type OauthServerKey = 'google' | 'kakao' | 'naver';
+let error = false;
+let isLogin = JSON.parse(localStorage.getItem('mockLogin') as string) || false;
 
 const getLoginUrl = rest.get(
   '/api/oauth-loginUrl/:oauthServer',
   async (req, res, ctx) => {
-    await sleep(1000);
     if (error) {
       error = false;
       return res(
-        // Send a valid HTTP status code
         ctx.status(403),
-        // And a response body, if necessary
         ctx.json({
           errorMessage: `url not found`,
         })
@@ -35,9 +33,11 @@ const getLoginUrl = rest.get(
     };
 
     if (oauthServer === 'google') {
-      await sleep(3000);
-    } else if (oauthServer === 'naver') {
       await sleep(1000);
+    } else if (oauthServer === 'naver') {
+      await sleep(2000);
+    } else if (oauthServer === 'kakao') {
+      await sleep(3000);
     }
 
     return res(ctx.json(oauthServerObj[oauthServer as OauthServerKey]));
@@ -46,6 +46,8 @@ const getLoginUrl = rest.get(
 
 const getAccessToken = rest.get('/api/oauth-login', async (req, res, ctx) => {
   const oauthCode = req.url.searchParams.get('code');
+  localStorage.setItem('mockLogin', 'true');
+  isLogin = true;
   await sleep(2000);
   return res(
     ctx.json({
@@ -78,37 +80,29 @@ const refreshAccessToken = rest.get(
 
 const checkRefreshToken = rest.get('/api/auth/check', async (req, res, ctx) => {
   await sleep(1000);
-  // return res(
-  //   ctx.json({
-  //     response: false,
-  //   })
-  // );
-  return res(ctx.json({ response: true }));
+  return res(ctx.json({ response: isLogin }));
 });
 
-// 만료 token 요청 api
-const test = rest.get('/api/test', async (req, res, ctx) => {
-  await sleep(2000);
-  return res(
-    ctx.json({
-      token_type: 'Bearer Token',
-      expires_in: 86400,
-      access_token:
-        'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4iLCJleHAiOjE1MTYyMzAwMH0.Q',
-      scope: 'photo offline_access',
-      refresh_token: 'k9ysLtnRzntzxJWeBfTOdPXE',
-    })
-  );
-});
-
-const test2 = rest.get('/api/test2', async (req, res, ctx) => {
-  // await sleep(2000);
-  return res(ctx.json({}));
-});
+const getExpiredAccessToken = rest.get(
+  '/api/expired-access-token',
+  async (req, res, ctx) => {
+    await sleep(2000);
+    return res(
+      ctx.json({
+        token_type: 'Bearer Token',
+        expires_in: 86400,
+        access_token:
+          'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4iLCJleHAiOjE1MTYyMzAwMH0.Q',
+        scope: 'photo offline_access',
+        refresh_token: 'k9ysLtnRzntzxJWeBfTOdPXE',
+      })
+    );
+  }
+);
 
 const logout = rest.get('/api/auth/logout', async (req, res, ctx) => {
-  // await sleep(2000);
-  // cookie에서 refresh token 삭제
+  isLogin = false;
+  localStorage.setItem('mockLogin', 'false');
   return res(ctx.json({}));
 });
 
@@ -116,8 +110,7 @@ const handlers = [
   getLoginUrl,
   getAccessToken,
   refreshAccessToken,
-  test,
-  test2,
+  getExpiredAccessToken,
   logout,
   checkRefreshToken,
 ];
