+++ src/pages/Home/index.tsx
@@ -1,38 +1,7 @@
-import HTTP from '@/api';
-import axios from '@/api/core';
-
 const Home = () => {
   return (
     <div>
       <h1>홈 페이지입니다.</h1>
-      <button
-        type="button"
-        onClick={() => {
-          // console.log('요청!');
-          axios.get('/test', { requireAuth: true });
-        }}
-      >
-        만료AT받기
-      </button>
-      <br />
-      <button
-        type="button"
-        onClick={() => {
-          // console.log('요청!');
-          axios.get('/test2', { requireAuth: true });
-        }}
-      >
-        버튼
-      </button>
-      <br />
-      <button
-        type="button"
-        onClick={() => {
-          HTTP.logout();
-        }}
-      >
-        로그아웃
-      </button>
     </div>
   );
 };
