+++ src/pages/TripList/index.tsx
@@ -1,13 +1,5 @@
-import styled from 'styled-components';
-
 const TripList = () => {
-  return <LoginButtonBox>TripList Page 입니다.</LoginButtonBox>;
+  return <div>TripList Page 입니다.</div>;
 };
 
 export default TripList;
-
-const LoginButtonBox = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
